requirements.txt
async-lru==1.0.2
pyaes==1.6.1
PyMySQL==1.0.2
Pyrogram==1.2.9
PySocks==1.7.1
python-dotenv==0.19.2
